Add user scope

+++ controllers/images.ts
@@ -6,6 +6,7 @@ import { s3Client, streamFileFromS3, deleteFileFromS3 } from '../fileStorage/s3'
 import { downloadLocalFile, removeLocalFile } from '../fileStorage/local'
 import { removeImageFiles } from '../fileStorage/common'
 import { defaultLimit } from '../config'
+import { getUserId } from '../utils/user'
 
 interface NewImage {
     _id?: string
@@ -20,18 +21,27 @@ export const upload_image = async (req: Request, res: Response) => {
 
     // TODO: Only allow images
 
+    let userId: string | undefined = undefined
+    if (process.env.IMAGE_SCOPE === 'user') {
+        const id = getUserId(res.locals.user)
+
+        // If no id found, throw an error
+        if (!id) throw createHttpError(401, 'User ID not provided')
+        userId = id
+    }
+
     const { originalname } = req.file
 
     const {
         _id,
         time,
-        file: userprovidedFilePath,
+        file: userProvidedFilePath,
         ...data
     } = parse_post_body(req.body)
 
-    const file = userprovidedFilePath || originalname
+    const file = userProvidedFilePath || originalname
 
-    const query = { file }
+    const query = { file, userId }
     const itemProperties: NewImage = {
         _id,
         time: time ? new Date(time) : new Date(),
@@ -59,6 +69,14 @@ export const read_images = async (req: Request, res: Response) => {
         select,
     } = parse_query(req.query)
 
+    if (process.env.IMAGE_SCOPE === 'user') {
+        const id = getUserId(res.locals.user)
+
+        if (!id) throw createHttpError(401, 'User ID not provided')
+
+        query.userId = id
+    }
+
     const items = await Image.find(query)
         .sort({ [sort]: order })
         .skip(Number(skip))
@@ -72,15 +90,41 @@ export const read_images = async (req: Request, res: Response) => {
 
 export const read_image = async (req: Request, res: Response) => {
     const { _id } = req.params
-    const image = await Image.findOne({ _id })
+
+    const query: Record<string, string> = { _id }
+
+    if (process.env.IMAGE_SCOPE === 'user') {
+        const id = getUserId(res.locals.user)
+
+        if (!id) throw createHttpError(401, 'User ID not provided')
+
+        query.userId = id
+    }
+
+    const image = await Image.findOne(query)
+
     if (!image) throw createHttpError(404, `Image ${_id} not found`)
+
     res.send(image)
 }
 
 export const read_image_file = async (req: Request, res: Response) => {
     const { _id } = req.params
-    const image = await Image.findOne({ _id })
+
+    const query: Record<string, string> = { _id }
+
+    if (process.env.IMAGE_SCOPE === 'user') {
+        const id = getUserId(res.locals.user)
+
+        if (!id) throw createHttpError(401, 'User ID not provided')
+
+        query.userId = id
+    }
+
+    const image = await Image.findOne(query)
+
     if (!image) throw createHttpError(404, `Image ${_id} not found`)
+
     const { file } = image
 
     if (s3Client) await streamFileFromS3(res, file)
@@ -91,6 +135,14 @@ export const update_images = async (req: Request, res: Response) => {
     const { query } = parse_query(req.query)
     const { $set, $unset } = parseUpdateBody(req.body)
 
+    if (process.env.IMAGE_SCOPE === 'user') {
+        const id = getUserId(res.locals.user)
+
+        if (!id) throw createHttpError(401, 'User ID not provided')
+
+        query.userId = id
+    }
+
     const result = await Image.updateMany(query, { $set, $unset })
 
     res.send(result)
@@ -100,8 +152,18 @@ export const update_image = async (req: Request, res: Response) => {
     const { _id } = req.params
     const { $set, $unset } = parseUpdateBody(req.body)
 
-    const updatedDoc = await Image.findByIdAndUpdate(
-        _id,
+    const query: Record<string, string> = { _id }
+
+    if (process.env.IMAGE_SCOPE === 'user') {
+        const id = getUserId(res.locals.user)
+
+        if (!id) throw createHttpError(401, 'User ID not provided')
+
+        query.userId = id
+    }
+
+    const updatedDoc = await Image.findOneAndUpdate(
+        query,
         { $set, $unset },
         { new: true }
     )
@@ -115,7 +177,18 @@ export const replace_image_data = async (req: Request, res: Response) => {
     const { _id } = req.params
     const data = req.body
 
-    const updated_image = await Image.findByIdAndUpdate(_id, { data })
+    const query: Record<string, string> = { _id }
+
+    if (process.env.IMAGE_SCOPE === 'user') {
+        const id = getUserId(res.locals.user)
+
+        if (!id) throw createHttpError(401, 'User ID not provided')
+
+        query.userId = id
+    }
+
+    const updated_image = await Image.findOneAndUpdate(query, { data })
+
     if (!updated_image) throw createHttpError(404, `Image ${_id} not found`)
 
     res.send(updated_image)
@@ -123,6 +196,14 @@ export const replace_image_data = async (req: Request, res: Response) => {
 
 export const delete_images = async (req: Request, res: Response) => {
     const { query } = parse_query(req.query)
+
+    if (process.env.IMAGE_SCOPE === 'user') {
+        const id = getUserId(res.locals.user)
+
+        if (!id) throw createHttpError(401, 'User ID not provided')
+
+        query.userId = id
+    }
 
     const items = await Image.find(query)
 
@@ -136,9 +217,23 @@ export const delete_images = async (req: Request, res: Response) => {
 
 export const delete_image = async (req: Request, res: Response) => {
     const { _id } = req.params
-    const image = await Image.findOneAndDelete({ _id })
+
+    const query: Record<string, string> = { _id }
+
+    if (process.env.IMAGE_SCOPE === 'user') {
+        const id = getUserId(res.locals.user)
+
+        if (!id) throw createHttpError(401, 'User ID not provided')
+
+        query.userId = id
+    }
+
+    const image = await Image.findOneAndDelete(query)
+
     if (!image) throw createHttpError(404, `Image ${_id} not found`)
+
     if (s3Client) await deleteFileFromS3(image.file)
     else removeLocalFile(image.file)
+
     res.send({ _id })
 }
